removes persistent question component, modifies question component
