--- v2/src/components/persistentQuestion/question.css
@@ -1,64 +1,0 @@
-.question-box {
-  width: 100%;
-  border-radius: .375rem;
-  border: 1px solid rgb(51, 51, 51);
-  display: flex;
-  flex-flow: column;
-  overflow: hidden;
-  margin-bottom: 20px;
-  box-shadow: var(--container-box-shadow);
-}
-
-.question-box-text {
-  padding: 1.3125rem 1.5rem;
-  font-size: 1.125rem;
-  color: white;
-  border-bottom: 1px solid rgb(51, 51, 51);
-  font-weight: bold;
-}
-
-.question-box-answers {
-  padding: .75rem 1.5rem 0rem;
-  display: flex;
-  flex-wrap: wrap;
-  position: relative;
-  background-color: rgb(42, 42, 42);
-}
-
-.question-box-answer {
-  display: block;
-  padding: 5px 20px;
-  background-color: rgb(54, 54, 54);
-  font-size: 1.125rem;
-  line-height: 22px;
-  font-weight: bold;
-  color: rgb(255, 153, 51);
-  border: 1px solid rgb(77, 77, 77);
-  border-radius: 6px;
-  margin-bottom: 12px;
-}
-
-.question-box-answer:not(:last-of-type) {
-  margin-right: 1.5rem;
-}
-
-.question-box-answer:hover {
-  cursor: pointer;
-  border-color: #f93;
-}
-
-.question-box-answer-slash {
-  color: rgb(255, 153, 51);
-}
-
-.question-box-submitted-container {
-  width: 100%;
-  display: flex;
-  border-radius: 6px;
-  background: rgb(42, 42, 42);
-  border: 1px solid rgb(51, 51, 51);
-  padding: 16px;
-  margin-bottom: 20px;
-  color: white;
-  box-shadow: var(--container-box-shadow);
-}
