+++ v2/thirdpartyemailpassword/quick-setup/frontend.mdx
@@ -9,7 +9,6 @@ import {Question, Answer}from "/src/components/question"
 import AppInfoForm from "/src/components/appInfoForm"
 import NpmOrScriptTabs from "/src/components/tabs/NpmOrScriptTabs"
 import RRDVersionSubTabs from "/src/components/tabs/RRDVersionSubTabs"
-import {PersistentQuestion, PersistentAnswer} from "/src/components/persistentQuestion"
 import AngularUIImplementation from "/src/components/reusableSnippets/angularUIImplementation"
 
 # Frontend Integration
